@@ -1,8 +1,8 @@
 module gitee.com/chunanyong/gowe
 
-go 1.12
+go 1.25
 
 require (
-	github.com/beevik/etree v1.1.0
-	golang.org/x/crypto v0.0.0-20210616213533-5ff15b29337e
+	github.com/beevik/etree v1.6.0
+	golang.org/x/crypto v0.42.0
 )
