@@ -1,13 +1,14 @@
 package gowe
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 )
 
-//WxMaCodeGetUnlimited 小程序码接口
-//https://developers.weixin.qq.com/miniprogram/dev/api-backend/open-api/qr-code/wxacode.getUnlimited.html
-func WxMaCodeGetUnlimited(wxMaConfig IWxMaConfig, body *WxMaCodeGetUnlimitedBody) ([]byte, *ResponseBase, error) {
+// WxMaCodeGetUnlimited 小程序码接口
+// https://developers.weixin.qq.com/miniprogram/dev/api-backend/open-api/qr-code/wxacode.getUnlimited.html
+func WxMaCodeGetUnlimited(ctx context.Context, wxMaConfig IWxMaConfig, body *WxMaCodeGetUnlimitedBody) ([]byte, *ResponseBase, error) {
 	apiurl := WxMpAPIURL + fmt.Sprintf("/wxa/getwxacodeunlimit?access_token=%s", wxMaConfig.GetAccessToken())
 	// 参数处理
 	bodyStr, err := json.Marshal(body)
@@ -25,7 +26,7 @@ func WxMaCodeGetUnlimited(wxMaConfig IWxMaConfig, body *WxMaCodeGetUnlimitedBody
 			"b": body.LineColorB,
 		}
 	}
-	data, err := httpPost(apiurl, params)
+	data, err := httpPost(ctx, apiurl, params)
 	// 发送请求
 	if err != nil {
 		return nil, nil, err
@@ -36,7 +37,7 @@ func WxMaCodeGetUnlimited(wxMaConfig IWxMaConfig, body *WxMaCodeGetUnlimitedBody
 	return data, baseErr, err
 }
 
-//WxMaCodeGetUnlimitedBody 小程序码请求参数
+// WxMaCodeGetUnlimitedBody 小程序码请求参数
 type WxMaCodeGetUnlimitedBody struct {
 	Scene      string `json:"scene"`                // 最大32个可见字符,只支持数字,大小写英文以及部分特殊字符:!#$&'()*+,/:;=?@-._~,其它字符请自行编码为合法字符(因不支持%,中文无法使用 urlencode 处理,请使用其他编码方式)
 	Page       string `json:"page,omitempty"`       // 必须是已经发布的小程序存在的页面(否则报错),例如 pages/index/index, 根路径前不要填加 /,不能携带参数(参数请放在scene字段里),如果不填写这个字段,默认跳主页面
