@@ -1,14 +1,15 @@
 package gowe
 
 import (
+	"context"
 	"encoding/xml"
 	"errors"
 )
 
-//WxPayDownloadBill 下载对账单 https://pay.weixin.qq.com/wiki/doc/api/jsapi.php?chapter=9_6
-func WxPayDownloadBill(wxPayConfig IWxPayConfig, body *WxPayDownloadBillBody) (*WxPayDownloadBillResponse, error) {
+// WxPayDownloadBill 下载对账单 https://pay.weixin.qq.com/wiki/doc/api/jsapi.php?chapter=9_6
+func WxPayDownloadBill(ctx context.Context, wxPayConfig IWxPayConfig, body *WxPayDownloadBillBody) (*WxPayDownloadBillResponse, error) {
 	// 业务逻辑
-	bytes, err := wxPayDoWeChat(wxPayConfig, "/pay/downloadbill", body, 0)
+	bytes, err := wxPayDoWeChat(ctx, wxPayConfig, "/pay/downloadbill", body, 0)
 	if err != nil {
 		return nil, err
 	}
@@ -23,7 +24,7 @@ func WxPayDownloadBill(wxPayConfig IWxPayConfig, body *WxPayDownloadBillBody) (*
 	}
 }
 
-//WxPayDownloadBillBody 下载对账单的参数
+// WxPayDownloadBillBody 下载对账单的参数
 type WxPayDownloadBillBody struct {
 	SignType string `json:"sign_type,omitempty"` // 签名类型,目前支持HMAC-SHA256和MD5,默认为MD5
 	BillDate string `json:"bill_date"`           // 下载对账单的日期,格式:20140603
@@ -31,7 +32,7 @@ type WxPayDownloadBillBody struct {
 	TarType  string `json:"tar_type,omitempty"`  // 非必传参数,固定值:GZIP,返回格式为.gzip的压缩包账单.不传则默认为数据流形式.
 }
 
-//WxPayDownloadBillResponse 下载对账单的返回值
+// WxPayDownloadBillResponse 下载对账单的返回值
 type WxPayDownloadBillResponse struct {
 	WxResponseModel
 	ErrCode string `xml:"err_code"` // 失败错误码,详见错误码列表 TODO
